requirements.txt
python-telegram-bot>=20.0
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv